+++ authentication/ldap.js
@@ -8,7 +8,8 @@ const LdapStrategy = require('passport-ldapauth')
   , config = require('../config.js')
   , userTransformer = require('../transformers/user')
   , AuthenticationConfiguration = require('../models/authenticationconfiguration')
-  , authenticationApiAppender = require('../utilities/authenticationApiAppender');
+  , authenticationApiAppender = require('../utilities/authenticationApiAppender')
+  , SecurePropertyAppender = require('../security/utilities/secure-property-appender');
 
 let authenticationOptions = {
 };
@@ -84,18 +85,20 @@ function doConfigure(passport, strategyConfig) {
 }
 
 function configure(passport, config) {
-
   if (config) {
-    doConfigure(passport, config);
+    SecurePropertyAppender.appendToConfig(config).then(appendedConfig => {
+      doConfigure(passport, appendedConfig);
+    });
   } else {
-    AuthenticationConfiguration.getConfigurationsByType('ldap').then(strategyConfigs => {
-      strategyConfigs.forEach(strategyConfig => {
-        if (strategyConfig) {
-          doConfigure(passport, strategyConfig);
-        }
-      });
+    AuthenticationConfiguration.getConfiguration('ldap', 'ldap').then(strategyConfig => {
+      if (strategyConfig) {
+        return SecurePropertyAppender.appendToConfig(strategyConfig);
+      }
+      return Promise.reject('LDAP not configured');
+    }).then(appendedConfig => {
+      doConfigure(passport, appendedConfig);
     }).catch(err => {
-      log.error(err);
+      log.info(err);
     });
   }
 }