+++ CHANGELOG.md
@@ -8,6 +8,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 ##### Features
 * Local user signup captcha
 * New authentication functionality ability under admin->settings (e.g. create, edit, etc.).
+* Moving security settings to more secure location.
 
 ##### Bug Fixes